// Types/Issue.ts
import { Discussion } from "./Discussion";
import { Label } from "./Label";

export type Issue = {
    assignees?: string[];
    author?: string;
    blocked?: boolean;
    blockedByCount?: number;
    blockingCount?: number;
    closedAsDuplicateOf?: string;
    closedAt?: string;
    commenters?: string[];
    createdAt?: string;
    description?: string;
    discussions?: Discussion[];
    downvotes?: number;
    dueDate?: string;
    hasEpic?: boolean;
    healthStatus?: string;
    humanTimeEstimate?: string;
    humanTotalTimeSpent?: string;
    iid?: string;
    iteration?: string;
    labels?: Label[];
    mergeRequestsCount?: number;
    moved?: boolean;
    movedTo?: string;
    participants?: string[];
    severity?: string;
    state?: string;
    taskCompletionStatus?: {
        completedCount: number;
        count: number;
    };
    timeEstimate?: string;
    title?: string;
    totalTimeSpent?: string;
    updatedAt?: string;
    upvotes?: number;
    userNotesCount?: number;
    weight: number;
    relatedMergeRequests?: string[];
};